Remove check for deleted vertices
The `"deleted"` key does not appear to be set on the Vertex (which also
does not implement the `get` method). This removes the check from the
zone population plugin, ideally with minimal consequences as this key
does not appear to be set or used in any of the primary class of model
components.